#8 [REMOVE] LazyVerticalGrid border 제거
