+++ core/src/main/java/com/sopt/core/type/CellType.kt
@@ -1,0 +1,5 @@
+package com.sopt.core.type
+
+enum class CellType {
+    Blank, DateHeader, TimeHeader, Data
+}
